+++ src/pages/RegisterPage.jsx
@@ -1,0 +1,14 @@
+export default function RegisterPage() {
+  return (
+    <div className="min-h-screen flex items-center justify-center bg-gray-50 py-12 px-4 sm:px-6 lg:px-8">
+      <div className="max-w-md w-full space-y-8">
+        <div className="text-center">
+          <h2 className="text-3xl font-bold text-gray-900">Register Page</h2>
+          <p className="mt-4 text-gray-600">
+            This page is being developed by Aman.
+          </p>
+        </div>
+      </div>
+    </div>
+  );
+}
